requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
markdown
pdfkit
